refactor(poc): AUROC and Learning curves implementation (#27)
This Pull request is for a clever refactoring of the previous k-NN implementation. I conducted a number of short studies with it and its documentation, and I determined that no hyper parameter other than 'k' is meaningful, which is why there is not advanced hyper parameter tuning in this PR. As a result, there is also an implementation of the learning curves from two different models. In addition, the AUROC curves were also implemented as a bonus of the task. Both evaluation metrics guarantee the viability of k-NN algorithm with our type of dataset. Please review the report on Awake's drive for any extra information; you will be able to examine the limitation as well as everything else needed to gain a thorough knowledge of this quick refactoring.

* refactor(poc): implement AUROC curves and generic plot learning validation curves
* fixup! refactor(poc): implement AUROC curves and generic plot learning validation curves